Rip out zoom logic and redo Graphics.drawScreen()
We're just trying to get to a functional drawScreen() now, though the rest of
the game is currently broken. Create a nice ASCII picture of the new playing
field setup.

Moved the last remnants of game setup out of the Model constructor and into
ModelFactory.createModel(). Move the functionality of Player.setX() into
Player.calcY().

Get rid of RunGameAct.onRestoreInstanceState() since it's never used.

+++ src/scorched/android/Weapon.java
@@ -69,7 +69,7 @@ public enum Weapon
             return true;
 
         // Use linear interpolation to check if we're under the ground
-        float h[] = model.getHeights();
+        short h[] = model.getHeights();
         int ix = (int)x;
         float xSlop = x - (float)ix;
         float interpolatedY = ((1.0f - xSlop) * h[ix]) + (xSlop * h[ix + 1]);
